Fixed tailwind, renamed awsClients to clients
