+++ packages/api/src/clients/aws/sesClient.ts
